Move events and WeightInfor trait to common; Add change_sigs_required_via_creator extrinsic
